Move teamMembersTable to containers
