+++ src/containers/WoffuBoard/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Header } from '../../components/header';
-import { TeamMembersTable } from '../../components/teamMembersTable';
+import { TeamMembersTable } from '../teamMembersTable';
 
 export const WoffuBoard = () => {
   return (
