+++ src/containers/teamMembersTable/cardRow/personalInfo/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Avatar } from '../../../avatar';
+import { Avatar } from '../../../../components/avatar';
 
 import { PersonalInfoWrapper, InfoWrapper, Name, JobInfo } from './styles';
